split each function into files
