+++ app.js
@@ -4,7 +4,7 @@ const cors = require('cors')
 const cookieParser = require('cookie-parser');
 require('dotenv').config();
 // const {validateBody} = require('./utils');
-const contactsRouter = require('./routes/api/contacts-routes')
+const router = require('./routes/api')
 
 const app = express()
 
@@ -16,7 +16,7 @@ app.use(express.json())
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 
-app.use('/api/contacts', contactsRouter)
+app.use('/api/contacts', router)
 
 app.use((req, res) => {
   res.status(404).json({ message: 'Not found' })
