+++ controllers/contacts/index.js
@@ -1,0 +1,17 @@
+const { ctrlWrapper} = require('../../utils');
+const listContacts = require('./listContacts');
+const getContactById = require('./getContactById');
+const addContact = require('./addContact');
+const updateContact = require('./updateContact');
+const updateStatusContact = require('./updateStatusContact');
+const removeContact = require('./removeContact');
+
+
+module.exports = {
+   listContacts: ctrlWrapper(listContacts),
+  getContactById: ctrlWrapper(getContactById),
+  addContact: ctrlWrapper(addContact),
+  updateContact: ctrlWrapper(updateContact),
+  updateStatusContact : ctrlWrapper(updateStatusContact ),
+  removeContact: ctrlWrapper(removeContact),
+}
