+++ controllers/contacts/listContacts.js
@@ -1,0 +1,8 @@
+const { Contact } = require('../../models');
+
+const listContacts = async (req, res) => {
+  const result = await Contact.find({})
+  res.json(result)
+}
+
+module.exports = listContacts;
