+++ controllers/contacts/updateStatusContact.js
@@ -1,0 +1,14 @@
+const { Contact }  = require('../../models');
+const { HttpError } = require('../../helpers');
+
+const updateStatusContact = async (req, res) => {
+    const { contactId } = req.params;
+    const { favorite } = req.body;
+    const result = await Contact.findByIdAndUpdate(contactId, { favorite }, { new: true });
+    if (!result) {
+        throw HttpError(404, `"Not found"`);
+    }
+    res.json(result);
+};
+
+module.exports = updateStatusContact;
