+++ controllers/index.js
@@ -1,15 +1,15 @@
 const { listContacts,
     getContactById,
     addContact,
-    removeContact,
     updateContact,
-updateStatusContact} = require("./contacts-controllers")
+    removeContact,
+    updateStatusContact} = require("./contacts")
 
 module.exports = {
     listContacts,
     getContactById,
     addContact,
-    removeContact,
     updateContact,
     updateStatusContact,
+    removeContact,
 }
